chore: dedupe gamepad adapter logic and fix camera resource leaks
Follow-up to #8/#11/#15's perf/redundancy cleanup passes.

- Extract the byte-for-byte identical get_state()/t() button-and-axis
  mapping shared by all 5 gamepad-to-Tello adapters (Xbox, Xbox One,
  GC102, Logitech F710, TectInter) into a shared BaseGamepadTelloAdapter,
  removing ~250 duplicated lines.
- mock_follow_face.py never released its webcam handle or destroyed the
  cv2 window on exit, unlike every other camera script in the repo; added
  the same cam.release()/cv2.destroyAllWindows() cleanup.
- follow_face.py never called cv2.destroyAllWindows() on exit, unlike
  sibling scripts; added it alongside the existing streamoff() cleanup.
- 3_drone_information.py imported time but never used it.
