+++ src/controller_adapters/xbox_one_tello_adapter.py
@@ -5,64 +5,15 @@ script_dir = os.path.dirname(__file__)
 parent_dir = os.path.join(script_dir, "..")
 sys.path.append(parent_dir)
 
-from typing import List
 from joysticks.pygame_connector import PyGameConnector
 from joysticks.xbox_one_controller import XboxOnePyGameController
-from services.tello_controller import (
-    TelloActionType,
-    TelloControlState,
-    TelloController,
-)
-
+from .base_gamepad_adapter import BaseGamepadTelloAdapter
 from .utils import run_adapter_test
 
 
-class XboxOneTelloControlAdapter(TelloController):
-
+class XboxOneTelloControlAdapter(BaseGamepadTelloAdapter):
     def __init__(self, controller: XboxOnePyGameController):
-        self.xbox_controller = controller
-
-    def t(self, controller_axis_value: float) -> int:
-        "Transform the controller axis value to the tello control value"
-        return int(controller_axis_value * 100)
-
-    def get_state(self) -> TelloControlState:
-        controller_state = self.xbox_controller.get_state()
-
-        pressed_buttons = controller_state.buttons.get_pressed_buttons()
-        d_pad = controller_state.d_pad
-
-        events: List[TelloActionType] = []
-        if "Y" in pressed_buttons:
-            events.append(TelloActionType.TAKEOFF)
-        if "A" in pressed_buttons:
-            events.append(TelloActionType.LAND)
-        if "B" in pressed_buttons:
-            events.append(TelloActionType.EMERGENCY_LAND)
-
-        if d_pad.vertical_up == 1:
-            events.append(TelloActionType.FLIP_FORWARD)
-        if d_pad.vertical_up == -1:
-            events.append(TelloActionType.FLIP_BACK)
-        if d_pad.horizontal_right == -1:
-            events.append(TelloActionType.FLIP_LEFT)
-        if d_pad.horizontal_right == 1:
-            events.append(TelloActionType.FLIP_RIGHT)
-
-        if "RB" in pressed_buttons:
-            events.append(TelloActionType.INCREASE_SPEED_CM_S)
-        if "LB" in pressed_buttons:
-            events.append(TelloActionType.DECREASE_SPEED_CM_S)
-
-        t = self.t
-
-        return TelloControlState(
-            forward_velocity=t(-controller_state.axes.left_stick.vertical_down),
-            right_velocity=t(controller_state.axes.left_stick.horizontal_right),
-            up_velocity=t(-controller_state.axes.right_stick.vertical_down),
-            yaw_right_velocity=t(controller_state.axes.right_stick.horizontal_right),
-            events=events,
-        )
+        super().__init__(controller)
 
 
 if __name__ == "__main__":
